test_install_receipts.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<window title="Mozilla Bug "
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="apphelper.js"/>
  <script type="application/javascript" src="jshelper.js"/>
  <script type="application/javascript" src="urlmatch.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id="
     target="_blank">Mozilla Bug </a>
  </body>

<script> 

steps = [setUp, install_super_crazy, tearDown];

runAll(steps);

function setUp(next) { 
  info("in " + arguments.callee.name);
  uninstallAll(next);
}


function install_super_crazy(next)  {
  info("in " + arguments.callee.name);
  var appURL = SERVERS['super_crazy'];
  popup_listener();
  var pending = navigator.mozApps.install(appURL, {receipts: ["a",0, true,false,null,undefined,{one:1},{nest:{ed:"hi"}},NaN,Infinity]});
  pending.onsuccess = function ()  {
    ok(pending.result.receipts[0] == "a", pending.result.receipts[0]);
    ok(pending.result.receipts[1] == 0, pending.result.receipts[1]);
    ok(pending.result.receipts[2] == true, pending.result.receipts[2]); 
    ok(pending.result.receipts[3] == false, pending.result.receipts[3]); 
    ok(pending.result.receipts[4] == null, pending.result.receipts[4]); 
    ok(pending.result.receipts[5] == undefined, pending.result.receipts[5]);
    ok(pending.result.receipts[6].one == 1, pending.result.receipts[6].one); 
    ok(pending.result.receipts[7].nest.ed == "hi", pending.result.receipts[7].nest.ed);
    todo(pending.result.receipts[8] == NaN, pending.result.receipts[8]);
    todo(pending.result.receipts[9] == Infinity, pending.result.receipts[9]);
 
    next();
  };
}

function tearDown() {
  info("in " + arguments.callee.name);
  uninstallAll();
}

</script> 

</window>
